+++ backend/src/main/java/com/digginroom/digginroom/oauth/jwk/KakaoJwkProvider.java
@@ -4,7 +4,7 @@ import com.auth0.jwk.Jwk;
 import com.auth0.jwk.JwkException;
 import com.auth0.jwk.JwkProvider;
 import com.auth0.jwk.JwkProviderBuilder;
-import com.digginroom.digginroom.domain.member.Provider;
+import com.digginroom.digginroom.domain.member.vo.Provider;
 import com.digginroom.digginroom.exception.OAuthResolverException;
 import org.springframework.stereotype.Component;
