+++ backend/src/test/java/com/digginroom/digginroom/controller/MemberArgumentResolverTest.java
@@ -17,7 +17,7 @@ class MemberArgumentResolverTest extends ControllerTest {
         RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
                 .when()
-                .get("/member/me")
+                .get("/member/genres")
                 .then()
                 .statusCode(HttpStatus.UNAUTHORIZED.value())
                 .body("timeStamp", Matchers.notNullValue())
@@ -31,7 +31,7 @@ class MemberArgumentResolverTest extends ControllerTest {
                 .cookie("JSESSIONID", "anything")
                 .contentType(ContentType.JSON)
                 .when()
-                .get("/member/me")
+                .get("/member/genres")
                 .then()
                 .statusCode(HttpStatus.UNAUTHORIZED.value())
                 .body("timeStamp", Matchers.notNullValue())
